Use view component
